+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Features
 - model picker no longer provider based, :NAIModel lets you choose any model,
   not just current provider
 - use snacks as default picker (if installed, then telescope or fzf)
-
+- tweaked config 
 
 # 1.0.0 (2025-04-17)
 Features